+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 - Create a list with `Iterable::Lists#create`
 - Delete a list with `Iterable::Lists#delete`
 - `/lists/getUsers` with `Iterable::Lists#users`
+- `/lists/subscribe` endpoint added as `Iterable::Lists#subscribe`
 
 ### MessageTypes